Print report to console if no output file provided
This commit changes Tern's reporting to only print output to a file if a
file is specified by the user using '-o <output file>'. If no output
file is specified, the report will be printed to the console. As a
result, we can delete the constants 'report.txt' and 'report.html'.

This commit also updates the README and Tern's CLI menu to inform users
that they need to provide an output file, otherwise the report will be
printed to the console. Additionally, a section on HTML reporting was
added to the README and all instances of BoM were changed to SBoM.

Signed-off-by: Rose Judge <rjudge@vmware.com>