Remove raise immediately statements
This commit removes raise only statements in try-except branches
found by pylint. Although the purpose was to re-raise the same
exception, it makes sense to be explicit about it. The change
basically replaces the empty raises with an explicit raising of
the same exception that the try branch brakes on.

Fixes #201

Signed-off-by: Nisha K <nishak@vmware.com>