Fix package version collection for Alpine
The apk command to collect version information for Alpine packages was
collecting the package name and version instead of just the isolated
version string. This was causing Tern's output file to report the
package as pkgname-pkgname-version. This commit fixes the version
command under the apk package manager section in base.yml to only
collect the package version information.

Resolves #758

Signed-off-by: Rose Judge <rjudge@vmware.com>