Amend Dockerfile and docker_run for fuse-overlayfs
Tern does not currently run within a container image without an external
mount to a Linux host. This commit updates the Dockerfile to install
fuse-overlayfs which will allow Tern to run in a container on a
non-Linux host. This commit also updates the Dockerfile used for
continuous integration testing. A few notable changes in the Dockerfile
include:

- Change base OS from photon to debian. This is a requirement of
fuse-overlayfs.
- Remove the creation of the hostmount directory, as it is no longer
required for mounting outside of the container.
- Removes the --quiet option in the ENTRYPOINT so the user can see the
default logging behavior.
- Adds the '--driver fuse' command line option to the ENTRYPOINT to
enable fuse-overlayfs as the default mounting driver.

This commit also updates the docker_run.sh script to no longer mount to
the hostmount directory and adds /dev/fuse as a host device to
enable fuse-overlayfs. test_files_touched.py is updated to
account for the changes in how the docker_run script should be invoked.
Lastly, there is an update to the README that reflects the changes to
commands required to run Tern in a Docker container.

Works towards #679

Signed-off-by: Rose Judge <rjudge@vmware.com>