Enable Terns run in a container for non-Linux host
Tern currently cannot run within a container image without an
external mount to a Linux host. This is due to the dependency on the
Linux filesystem's overlayfs driver that is used to mount the
layers which does not natively exist on Mac or Windows. This commit
makes changes to enable Tern to run on a non-Linux host by using the
FUSE implementation for overlayfs when Tern is being run in a container.

First, this commit deprecates the --bind-mount CLI option. Instead, a
--driver CLI option is added that will use the FUSE implementation of
overlayfs to do any mounting in a container, regardless of the host OS.

Since the --bind-mount CLI option is now deprecated, this commit also
makes the necessary updates in cache.py to any mount_dir references and
instead uses the correct working_dir.

Resolves #679

Signed-off-by: Rose Judge <rjudge@vmware.com>