Add support to execute commands from WORKDIR
Currently tern does not track workdir while executing
base. This becomes very necessary while inspecting golang
images because the scripts for collecting metadata for
go modules rely on the files present in workdir.

This PR adds functionality to execute the scripts in the
workdir in which they need to be executed. It is achieved
by looking for workdir in each layer. If their is no
workdir then the commands are executed in previous layers
workdir.

A function is added in classes/docker_image.py to get the
final workdir. It is utilised by verify_invoke script.

A function is added in classes/image_layer.py to return
current layers workdir.

In analyze/docker/analyze.py changes are made to collect
workdir which then passes through a series of fucntion
to reach command_lib.py get_package_attr_list(). Where
cd workdir command is added to snippet_list.

Resolves #766

Signed-off-by: Abhay Katheria <abhay.katheria1998@gmail.com>