Update utilities to set working directory
The set_mount_dir() function was deprecated in place of
set_working_dir() when the --bind-mount CLI option was also deprecated.
This commit updates the container_debug.py and verify_invoke.py debug
scripts to make the right function call as well as the
test_class_docker_image.py file.

Signed-off-by: Rose Judge <rjudge@vmware.com>