Set environment vars before package collection
Some Docker images have environment variables defined in the config.json
file. This is especially true in images where go modules are used
or installed as this is where the PATH and GOPATH are set. This
commit adds functionality to set/export the environment variables from
the config in each image layer before package information is collected.
If environment variables exist in the image config that are not relevant
to the layer and its method of package collection, the environment
variables will simply be ignored by the package collection commands.

Works towards #695

Signed-off-by: Rose Judge <rjudge@vmware.com>