Enable golang package listing
This commit adds commands in snippets.yml and base.yml to
to capture golang module name, version, proj_url and copyright
metadata in the output report. This commit does not currently
collect license information for golang modules as more advanced
parsing is required but functionality for this will come in
the future.

Work towards #695

Signed-off-by: Abhay Katheria <abhay.katheria1998@gmail.com>