Add logic to find shell in Alpine images
Alpine images symlink the /bin/sh utility to an absolute path,
/bin/busybox. The current get_shell() function uses os.path.exists to
verify the existence of the shell path. In the case of Alpine images,
however, os.path.exists follows the /bin/sh symlink to the root of the
host container, /bin/busybox, which will only exist if busybox is
installed on the host container. This commit adds logic to follow
absolute shell path symlinks to the correct binary location in the
mounted working directory instead of looking for it in the root of the
host container.

Resolves #769

Signed-off-by: Rose Judge <rjudge@vmware.com>