Refactor Dockerfiles
The Dockerfiles were previously quite hard to read as the commands
were done as one-liners. This change adds newlines to the commands
to improve the readability.

Installation of pip packages is also moved to its own layer, which
allows us to utilize the layer cache for apt packages in case more pip
packages need to be installed later on (e.g. scancode-toolkit) to the
image.

Signed-off-by: Isac Sund <isac@isacsund.com>