use `python:3.7-slim` base image and install local copy of tern
This commit uses the official python Docker image to reduce the number of packages to install with `apt-get`. This also greatly decreases the final image size, by about 400 MB. This commit also installs the local copy of tern. This is useful because if a developer makes changes to tern, it is possible to test those changes with tern's docker image. Before this commit, tern was installed from PyPI, so even if changes were made to the source, the latest PyPI version would be installed.