load: New code section for external interactions
This is work towards #802

This commit creates a place for code associated with downloading
container images and creating disk layouts for them that follow
the two currently widely used container image schemas. Docker uses
schema version 1 to save images. Furthermore, Dockerfiles are a
Docker specific artifact, so it makes sense to separate those
operations from the rest of the code.

This commit does the following:
- Copy docker build and save functions from
tern/analyze/docker/container.py.
- Modify the functions so they are testable by themselves.
- Added tests for some of the functions. This should help with
  code coverage. To this end, added a couple more test fixtures and
  a passing and failing Dockerfiles.
- Added the new section and associated test to the test suite for
  CI.

Signed-off-by: Nisha K <nishak@vmware.com>