classes: Use load functions in DockerImage
This is work towards #802

Now that we have a module dedicated to translating image and tags
into an on-disk layout, we will move all the required functions
for backwards compatibility. The first step is to modify the
DockerImage class to be able to make use of these new functions
and separate the extraction of the image from the loading of the
metadata itself. Architecturally, the classes are only meant to
carry all the metadata for the image and not do any image operations
itself.

- Moved some required functions from container.py to docker_api.py
  for backwards compatibility. They will be removed in the future.
- Modified the DockerImage class to use these new functions.
- Removed the extraction of the docker container image from the
  class as that operation should be dependent on the command line
  arguments.
- Modified the tests to use the fixtures such that the docker image
  dump happens before loading of the image.

This commit will break the build.

Signed-off-by: Nisha K <nishak@vmware.com>