Hook up docker_api to setup and teardown
This is work towards #801 and #802

In order to move forward with code re-organization, we move
setup and teardown functions from report.py to the new file
prep.py. We then replace all the functions in the main execution
path for Docker images with the functions from prep.py and
docker_api.py. For the replacement, we needed to make changes
to rootfs's extract_tarfile to return success or failure.

This commit will fix the typical execution path for docker images
but breaks for other functions.

Fixes #808

Signed-off-by: Nisha K <nishak@vmware.com>