Fix loading package files from cache
Now that we have scripts that will also inventory files in packages,
we no longer have to re-read file data at the layer level when there
is a cache hit. We just load any files that are associated with the
layer.

- Load files and packages if there is a cache hit.
- If there is a cache miss, read file names, paths, and checksums
  at the layer level.

In order to make the file level data consistent between loading from
the cache and saving to the cache, we have to create FileData
dictionaries after inventorying package level files.

- Modify collate_package_metadata to convert the files information
  for each package into a list, thus creating a list of lists. The
  outer list corresponds to the number of packages found after
  inventorying. The length of the outer list should be the same as
  the number of package names.
- Modify convert_to_pkg_dicts to convert the file path lists under
  package inventory into FileData dictionaries for easy saving to
  the cache.
- Modify the fill method in the Package class to convert the files
  list into a list of FileData objects.

Signed-off-by: Nisha K <nishak@vmware.com>