Fix checksum parsing and Dockerfile building
This commit adds functionality to docker_api.py to build a
container image from a Dockerfile using the Docker API. Ideally
this is the only reason to use the Docker API.

In order to switch over to the new Docker API functions, we need
some way to pass the checksum type and checksum when reading the
extracted image after the image is removed and the client is
closed. To allow for this, we modify the dump_docker_image and
build_and_dump functions to return some image metadata, which we
then pass to the init functions for the DockerImage class. In this
way, we won't need to open up a client again. This also allows us
to move the Dockerfile report and lock execution to use the new
functions.

- Added build_and_dump function to build and extract a Docker image.
- Modified the dump_docker_image function to return image metadata
  instead of True or False.
- Modified the DockerImage constructor to take an optional checksum
  string of the form <digest_type>:<digest>. Updated the unit tests
  accordingly.
- Updated the Docker image and Dockerfile execution subroutines.

Signed-off-by: Nisha K <nishak@vmware.com>