Remove container.py and some deprecated functions
This completes the work towards #802

This commit removes container.py. Now all Docker API operations
use docker_api.py and we have a location that will be used for
external interactions with the general container ecosystem such
as container registries or client tools. The commit also deprecates
invoking scripts in a Docker container as the tool has been using
the bare minimum unsharing of the process and invoking scripts
in a chroot jail.

- Remove file container.py.
- Remove is_build function in favor of build_and_dump_image function
  in docker_api.py
- Remove invoke_in_container function. Now we exclusively use
  invoke_in_chroot. Thus we don't need chroot=True and override
  arguments in get_pkg_attr_list.
- Modified get_docker_image_digest to open up a client, download
  an image, get the digest, and close the client.
- Remove setup and teardown functions from report.py in favor of
  setup and teardown functions in prep.py

Signed-off-by: Nisha K <nishak@vmware.com>