Move command_lib into default and organize
This is work towards #803

The command_lib is now part of Tern's default operation. Therefore
its logical location is in the default folder. It also has some
operations that belong in some other function sets, such as
collecting the results of invoking retrieval commands from the
command library.

We add some new python files to separate the collection of the
data after invoking the commands and the filtering of any commands
found in the created_by image metadata as they are not necessarily
related to each other. This separation should take care of adding
a backup operation if there is no created_by data in the image
metadata.

This commit will break operation. Do not apply without the rest of
the commits in the chain

Signed-off-by: Nisha K <nishak@vmware.com>