Fix single layer analysis
This is work towards #803

This commit adds changes that will make single layer default
analysis work. The changes primarily contain the addition of
execute_base function which is the core execution for analyzing
images with one layer. The changes also clean up unneeded
execution steps. For example, we no longer perform any mounting
if we are get a cache hit.

Other changes are for the execution to complete successfully with
and without a cache hit.

Multi-layer execution is still broken.

Signed-off-by: Nisha K <nishak@vmware.com>