Fix multi-layer container analysis
This is work towards #803

This commit fixes analysis of container images having more than
one layer. Some changes to the core operation were required to
allow for environment variables to be passed to execution of
the base scripts. There is also a change to the collection of
metadata in an image's created_by key if the key exists in the
config. This allows for changes to be added to deal with situations
where there is no information about what script created the layer.
It is expected that get_commands_from_metadata will deal with
various possibilities while processing of these possiblities can
exist in other locations in the code.

Dockerfile analysis and lock functionality is still broken.

Fixes #789

Signed-off-by: Nisha K <nishak@vmware.com>