Complete Dockerfile analysis
This is work towards #803

This commit fixes execution for Dockerfile analysis. The Dockerfile
specific functions are roughly divided into the parse and lock
functions. This still needs some clean-up. run.py contains the
entrypoint. This commit also removes a stray print statement
in the Command class.

Comprehensive testing is still not complete

Signed-off-by: Nisha K <nishak@vmware.com>