Shorten fill_package_metadata function
This is work towards #803

fill_package_metadata is currently not used as the command library
doesn't have listings for non package manager installation methods.
It is not removed because it may come in handy in the future.
However, it still has too many branches to pass linting tests.
Hence this commit reduces the complexity of the function.

Signed-off-by: Nisha K <nishak@vmware.com>