Fix unbound local error when repo digest is given
The DockerImage class can accept an image string in the form
image:tag or image@digest_type:digest. It can also accept an
optional digest string of the form digest_type:digest if the image
string is not known.

For the current status of running analysis on Docker style container
image names, the optional digest is not required. Therefore this
commit removes using the second digest string. To that effect
the changes are made in the extract_image, load_full_image and
load_base_image functions.

This also fixes #797

Signed-off-by: Nisha K <nishak@vmware.com>