Fix Dockerfile build with context
The docker module does not support inputting the context along
with the Dockerfile and assumes you are calling the script from the
same directory as the Dockerfile is located. This commit uses
a hack to send the context path with the Dockerfile contents. This
method only works when passing the Dockerfile contents rather than
the file, so we read the file contents first. We also rename some
variables to avoid confusion with the lambda variables being passed

Signed-off-by: Nisha K <nishak@vmware.com>