Add subcommand debug to replace verify_invoke
This is work towards #660

"Debug" is a loose term with regards to this kind of operation.
We're basically "stepping-though" the layers of the container
image or running scripts in an already mounted environment. This
subcommand covers operations that will be useful for developers
of Tern to troubleshoot any issues with the default operation.

The entrypoint is in tern/analyze/default/debug/run.py. This
is does not use a logger and is meant to be used in a development
environment where stdout is available to look at.

We move verify_invoke.py into a folder called "debug" under
the default operation where "debug" will contain all the operations
related to stepping through a container image.

We then edit run.py to use the existing module paths. Finally, we
add the subcommand "debug" to __main__.py and check if the user
has selected it in do_main.

Signed-off-by: Nisha K <nishak@vmware.com>