Move container_debug script into debug subcommand
This is work towards #660

This commit moves the container_debug.py script into the debug
subcommand. We add a '--step' option to go into interactive mode
where we ask what layer the user wants to inspect. When the user
chooses the layer, Tern will mount upto the layer selected and
provides a command to chroot into the layer to look around. We
also roll in the '--recover' option to unmount the layers after
the user is done inspecting.

We can now remove the container_debug.py script

Signed-off-by: Nisha K <nishak@vmware.com>