Use LicenseRef for spdxjson File licenses
Commit bca41a2 added support to generate SPDX JSON documents. In this
implementation, however, the file-level licenses found in the image are
reported in the document in plain text, which results in a validation
error in the SPDX tooling. In order to fix the validation error and get
the SPDX JSON document to validate, the File LicenseRef should be used
instead of the plain text license.

This commit changes the "licenseInfoInFiles" and "licenseInfoFromFiles"
SPDX JSON document information to list the File's SPDX LicenseRef
instead of the plain text license information associated with the file.

Signed-off-by: Rose Judge <rjudge@vmware.com>