Add new function get_extracted_text_dict()
This commit adds a new function, get_extracted_text_dict(), to
spdxjson/formats.py. This new function takes in a plain text license as
well as an associated license_ref and returns a formatted dictionary
that can be inserted in to an SPDX JSON document. This function is
required in preparation for adding a "hasExtractedLicensingInfos"
entry to the SPDX JSON document.

Signed-off-by: Rose Judge <rjudge@vmware.com>