Add extractedText licenseRef info to spdxjson docs
A previous commit, bca41a2, added support for a SPDX JSON format.
While that commit *does* produce technically valid JSON documents,
it does not include "extractedText" information for the variuos
LicenceRefs within the document. Without a list of extractedText
license information a user would have no way of knowing what the plain
text of a LicenseRef actually is. This commit fixes that and adds the
"hasExtractedLicensingInfos" entry to the SPDX JSON document that Tern
creates.

Signed-off-by: Rose Judge <rjudge@vmware.com>