Set spdxjson image downloadLocation to NOASSERTION
Currently, the spdxjson format uses the image repotag as the
downloadLocation value when producing SPDX documents. This is not
correct in accordance with the 2.2 spec, however. This commit sets the
downloadLocation value to 'NOASSERTION' for image 'packages'.

This commit also sets the downloadLocation to 'NONE' for image layer
'packages' instead of using the location of the tarball in the image.

As a result of the new downloadLocation values, there is some
cleanup of function parameters and function descriptions.

Works towards #855

Signed-off-by: Rose Judge <rjudge@vmware.com>