Set spdxtagvalue PackageDownloadLocation to NONE
Currently, the spdxtagvalue format uses the image repotag as the
PackageDownloadLocation value when producing SPDX documents. This is
not correct in accordance with the 2.2 spec, however, as the spec
requires that the value must match a certain pattern starting with
git|hg|svn|bzr or https://www or use NONE/NOASSERTION.

This commit sets the PackageDownloadLocation to NOASSERTION for
image 'packages' and NONE for image layer 'packages' instead of
using the repotag and location of the tarball in the image,
respectively.

Works towards #855

Signed-off-by: Rose Judge <rjudge@vmware.com>