Make scancode use more processes
Use len(os.sched_getaffinity(0)) to respect possible restrictions set on
CPU usage. User can limit the CPU usage of scancode with taskset on
linux as scancode will use all available cores.

The downside of this is that it is UNIX only approach, yet using
cross-platform approach would need additional dependency- the most
popular being psutil.

If during configuration time, when Python core is built, compilation of
code with sched_setaffinity will fail and HAVE_SCHED_SETAFFINITY is not
set, the call will not be avaliable and tern should fall-back to calling
scancode without -n option to respect it's defaults.

Signed-off-by: Alexander Mazuruk <a.mazuruk@samsung.com>