Fix tests for human readable id creation
The Image class now takes a repotag string of the form:
'image:tag' or 'image@hash:digest'. This commit updates the
documentation for the class and the corresponding test.

Since the human readable id is based on repo name and a repo tag,
we update the tests to check if the appropriate string is returned
For different combination of settings.

Signed-off-by: Nisha K <nishak@vmware.com>