Refactor get_os_style to set the right properties
This commit refactors the get_os_style function to set OS and
package format information in the ImageLayer instance so that the
information is available to all formats and not just stored in
the notices. To simplify the implementation, the package format
was removed from the OS reporting formatted string.

Signed-off-by: Nisha K <nishak@vmware.com>