Fix duplicate package reporting for alpine images
The apk package version collection utility in base.yml was not correctly
parsing version strings for the "libstdc++" package name (and possibly
others with non-alpha numeric characters). This was causing Tern to
report an empty version string for the libstdc++ package which led to
missing version strings and incorrect layer packages in the output
report. This commit slightly tweaks the package version collection
script to use sed instead of awk in order to properly collect and
report the package version information.

Resolves #881

Signed-off-by: Rose Judge <rjudge@vmware.com>