Fix error msg when no created_by info for layers
Squashed images do not contain created_by information for layers.
Running Tern on a squashed image exposed a bug where Tern tries
to report on the created_by information of a layer even though the
information is not available. To resolve this fix, this commit adds a
new error message, no_layer_created_by, that alerts the user that
layer creation info is not available and uses this in the output
report.

Resolves #838

Signed-off-by: Rose Judge <rjudge@vmware.com>