Refactor: pass arguments as a single object
To reduce code complexity, we introduce the concept of "options"
and "prerequisites". Options are passed as command line options.
Prerequisites are required arguments to a function. This is a
new class introduced in core.py and used to pass arguments to
execute_base and execute_snippets. The way we do it is create
the prereqs object in analyze_first_layer and
analyze_subsequent_layers. This reduces the complexity of passing
too many arguments to these functions.

Further, we follow this pattern in the dockerfile execution path.

Fixes #868

Signed-off-by: Nisha K <nishak@vmware.com>