Add GitPython as a runtime dependency
GitPython is currently a development dependency but when e01f826 was
merged it became a runtime dependency as well. This commit adds
GitPython to requirements.txt and removes it from tox.ini, setup.cfg
and pull_request.yml, which are all files where it was referenced
explicitly as a development dependency.

Resolves #917

Signed-off-by: Rose Judge <rjudge@vmware.com>