Refactor: Combine distroless into generic "host"
The "host" method should be available to all OSs that don't have
a shell and need a host script. This commit refactors the
"distroless" code such that the code execution remains OS agnostic.
Enabling new OSs should be done in the command_library and not
in the code.

- First we add some new properties to the Prereqs object that
  pertain to invoking a host script. That is a host path and a
  host shell.
- Next we refactor the core execution path to take just a Prereqs
  object rather than individual variables.
- Then we refactor the single_layer and multi_layer execution to
  set properties in the Prereqs object that is passed to it.
- We finally remove the duplicated code pertaining to "distroless"

This commit should also take care of the situation where there
is no commands or unrecognized commands in the image history. In
this case, there will be a fallback to executing the base script
entry.

Signed-off-by: Nisha K <nishak@vmware.com>