Add jq dependency to Dockerfiles and README
Commit e0785a0 introduced new functionality that uses the host system's
jq package utility to detect and parse npm package metadata. This commit
adds jq as a dependency to the Tern Dockerfiles and also updates the CI
tests to add a npm-based docker image, node:12.16-alpine, as a test
image. Finally, this PR updates the README to include jq as a host
system dependency.

Signed-off-by: Rose Judge <rjudge@vmware.com>