bugfix: Escape empty braces when parsing snippets
An issue was found in Tern when empty curly braces were present in a
Dockerfile command (i.e. ENV var=AUTH_{}). Python was interpreting these
empty curly braces as a string to be formatted and thus, throwing a
ValueError. This commit fixes this bug by adding escape braces to the
string in the collate_snippets() function so that Python does not try to
interpret them as a string to be formatted.

Resolves #913

Signed-off-by: Rose Judge <rjudge@vmware.com>