Set WORKDIR directory to collect go modules
The WORKDIR command in a Dockerfile is used to define the working
directory of a container which means that any RUN, CMD, ADD, COPY or
ENTRYPOINT command that follows a WORKDIR command will be executed in
the working directory.

PR #768 added support to set the working directory which is required
especially for go module metadata collection. Part of this work was lost
in various refactors so this commit adds back the necessary code to set
the working directory properly. It also removes an attempt to set the
working directory in the base layer as the WORKDIR will always be set
after the first layer. As a result of this commit, Tern properly
collects go module information.

Resolves #924

Signed-off-by: Rose Judge <rjudge@vmware.com>