fix: Include information about layer's created_by
When refactoring the default execution path into single_layer
and multi_layer, the info notice about the layer's created_by
was skipped. Granted, this information is probably repeated in
the layer object anyway, but this is the simplest way to add it
back to the reports because that information may or may not
be relevant to the consumers of the reports.

Signed-off-by: Nisha K <nishak@vmware.com>