SPDX JSON SBoM generation at container build time
This commit completes work on #850

We enable the SPDX JSON format for each layer snapshot at container
build time. At this point of container creation, we do not know the
nature of the package that contains all the files that are created
and software packages that are installed. Hence the document only
describes the packages and files. The generation of the SPDX
documents that reference these documents is out of scope for this
tool.

We will also tombstone the SPDX tagvalue format for now. At this
time, its use may not be relevant for the container build use
case.

Signed-off-by: Nisha K <nishak@vmware.com>