Deprecate command library commands in reports
The reports are becoming too verbose when printing out in detail the
commands that were used to fetch the metadata information. The logging
already has this information. In order to simplify the reports, this
commit removes adding notices about collection information to the layer
origin object and, as a result, deprecates the print_base_invoke()
function.

Resolves #823

Signed-off-by: Rose Judge <rjudge@vmware.com>