CI: Test lock with single stage Dockerfile
At this time "lock" does not work well with multistage Dockerfiles
especially when using the "as" clause to alias build stages. As
a result of switching to multistage Dockerfiles for building
Tern's own container images, the test that uses these Dockerfiles
fail. We replace the test Dockerfile with an old copy of Tern's
Dockerfile.

- Add Tern's old Dockerfile to samples
- Change tests to use this Dockerfile

Signed-off-by: Nisha K <nishak@vmware.com>