Better parsing of created_by values
When using Docker or Buildkit to build container images,
the config will contain a `created_by` key which contains the
shell command that was run to create the container image layer.
Typically, this command is followed by a comment #(nop) indicating
the commented out part of the shell command. This applies to
Docker directives other than RUN. We want to be able to not only
extract the RUN command but also if directives like ADD and COPY
were used to indicate that the added content is unknown in the
reports.

- Modified the get_run_command function to return the command and
  the instruction directive used from the created_by line.
- Modified the get_commands_from_metadata function to use the tuple
  returned by the get_run_command function to check for ADD and
  COPY instructions.
- Moved error message for no created by information from
  report/formats.py to report/errors.py as it is an error message.
- Modified the lock_dockerfile function to use the tuple returned
  by the get_run_command function to match the RUN commands in the
  image to the RUN instructions in the Dockerfile.

Fixes #918

Signed-off-by: Nisha K <nishak@vmware.com>