Added copyright info for NPM packages
Copyrights are extracted by checking if common licensing file names
exists in the package and if they are found the content is then
returned.

As the scanning is split between container and host (due to the jq
tool availability) the package path is left trimmed to remove first
'/' - effectively converting it to relative path which can then be
resolved correctly from the current working directory of host
command.

More license files to check can be added in the future by extending
the "files" variable

Resolves: #958
Signed-off-by: Matej Zachar <mzachar@simpleway.cz>