Make "tern lock" work for multistage docker file
Problem:
"tern lock" was not able to analyze multistage
docker files properly due to less understanding
of docker file's ".. AS .." syntax and not able
to relate images with each other.

Solution:
This commit tries to solve this problem. It uses
previous image's "import_str" property and set
it for the current image if current image's
"import_str" is not set.

Resolves: #951, #969

Signed-off-by: Mukul Taneja <mtaneja@vmware.com>