Add 'Understanding the Reports' section to README
Reports of tern provides a lot of information, and the types and format
of the items provided varies depending on the report format. Thus,
users might not understand what kind information they can get from
the reports and where to look at to get the information on the report.
This commit adds information about the specific type of information
available in each report format to the README.

Resolves #960

Signed-off-by: Rose Judge <rjudge@vmware.com>