formats: Add consumer abstract base class
This is work towards #946

We introduce the concept of a "consumer" which will convert
one or more reports of the same plugin type into an image object
or a list of image layers that can be manipulated later. This
addresses combining reports from another system in the same format
with reports generated by Tern, or working with reports generated
by Tern in a different layer context at container build time.

This commit adds an abstract base class called "Consume" under the
formats plugin machinery. It follows the same pattern as the
"Generate" abstact base class.

Signed-off-by: Nisha K <nishak@vmware.com>