formats: Add layer level JSON consumer
This is work towards #946

Using the Consumer abstract base class, we create a Consumer
class for the JSON format plugin that will create a list of
ImageLayer objects from the report. This is assuming the
report was created at the layer level, as is the case when
using the --live option.

As part the --live option, one can provide one or more reports
created from previous runs and the plugin will return a corresponding
internal object representation of those reports.

The implementation just uses the already existing "fill" methods
of the FileData and Package class to fill in data from the dictionary.

The layer objects are ordered based on the order of the reports in the
reports list.

Signed-off-by: Nisha K <nishak@vmware.com>