Record and report scancode file licenses
Running Tern with the scancode extension was not reporting file level
license information due to a logic error. This commit updates the
add_file_data() function in scancode/executor.py to update file metadata
from scancode if the file already exists in the cache, or if not, add
the file object to the layer file list.

This commit also changes the default format to report the name of the
file licenses instead of the license expression as the license is human
readable and consistent with the other human readable license
expressions typically reported by the package managers.

Resolves #959

Signed-off-by: Rose Judge <rjudge@vmware.com>